% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Project 1.R
\name{roll}
\alias{roll}
\title{Rolling Dice}
\usage{
roll(dietype = 1:6, probvalues = rep(1/length(dietype), length(dietype)))
}
\arguments{
\item{dietype}{The type of die you are rolling.}

\item{probvalues}{the probability of getting a praticular number of divots.}
}
\value{
the sum of two dice divots
}
\description{
Rolling Dice
}
\examples{
set.seed(123)
roll()
roll(1:20)
}

