% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Project 1.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle}
\usage{
shuffle(cards)
}
\arguments{
\item{Cards}{in a deck.}
}
\value{
Shuffels a deck of cards.
}
\description{
Shuffle
}
\examples{
deck <- shuffle(cards = deck)
deck[1:5, ]
}

