% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Project 1.R
\name{SQ}
\alias{SQ}
\title{Square Numbers}
\usage{
SQ(x)
}
\arguments{
\item{x}{must be a numeric object}
}
\value{
squares the value in the numeric object
}
\description{
Square Numbers
}
\examples{
sq(c(1,3,5))
sq(1:10)
}

